+++ backend/src/utils/schemas.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+from datetime import datetime
+
+
+class UserData(BaseModel):
+    name: str
+    email: str
+    channel_id: str
+    access_token: str
+    refresh_token: str
+    token_expiry: datetime
